Add modinterval package.
Package description:

Package modinterval provides data structures and functions for working with
1-dimensional integer intervals that use modular arithmetic.

See https://fgiesen.wordpress.com/2015/09/24/intervals-in-modular-arithmetic/
for a discussion of intervals in modular arithmetic.

The core type exported by the library is IntInterval, which is aware of its
modulus and has several methods for set-like operations, like Contains and
Expand*.

IntIntervals can also be transformed into standard real-number-based
intervals; see the RealIntervals() method. An implementation of such
intervals is provided in this package as well.